pasta1_etl/requirements.txt
boto3==1.35.82
botocore==1.35.82
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
idna==3.10
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==3.0.2
moto==5.0.23
py4j==0.10.9.7
pycparser==2.22
pyspark==3.5.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
responses==0.25.3
s3transfer==0.10.4
six==1.17.0
urllib3==1.25.10
Werkzeug==3.1.3
xmltodict==0.14.2
